Add assets for building docker images
The plan here is to deploy an ara installation to gke.
Before we can deploy it, we need to make some images, because ara
do not public images.

For ara, since it's a django app, use the opendevorg python-builder
and uwsgi-base images along with bindep so that we can produce
reasonably sized images with minimal effort.

For ara-web, use NodeJS 12 for the builder image and then copy that
to an Apache image for actually serving the build content.

In both cases we're bind-mount config into the running container.